CST-20334 fix delete query with dspaceobject

+++ dspace-api/src/main/java/org/dspace/app/metrics/dao/CrisMetricsDAOImpl.java
@@ -72,10 +72,13 @@ public class CrisMetricsDAOImpl extends AbstractHibernateDAO<CrisMetrics> implem
 
     @Override
     public void deleteByDSO(Context context, DSpaceObject dSpaceObject) throws SQLException {
-        String hqlQuery = "delete from " + CrisMetrics.class.getSimpleName() + " where resource=:resource";
-        Query query = createQuery(context, hqlQuery);
-        query.setParameter("resource", dSpaceObject);
-        query.executeUpdate();
+        if (dSpaceObject != null) {
+            UUID uuid = dSpaceObject.getID();
+            String hqlQuery = "delete from " + CrisMetrics.class.getSimpleName() + " where resource=:resource";
+            Query query = createQuery(context, hqlQuery);
+            query.setParameter("resource", uuid);
+            query.executeUpdate();
+        }
     }
 
     @Override
